+++ src/components/icon-box/index.ts
@@ -1,4 +1,5 @@
 export {TransportationIconBox} from './TransportationIconBox';
+export {TransportationIconBoxList} from './TransportationIconBoxList';
 export type {AnyMode, AnySubMode} from './types';
 export {getTransportModeSvg} from './utils';
 export {CounterIconBox} from './CounterIconBox';
